crypto-rebalancer/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crypto-rebalancer"
version = "0.1.0"
description = "Threshold-based Binance Spot portfolio rebalancing bot."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "python-binance>=1.0.19,<2.0",
    "pydantic>=2.5,<3.0",
    "PyYAML>=6.0",
    "APScheduler>=3.10,<4.0",
    "python-dotenv>=1.0",
    "typer>=0.12",
    "tenacity>=8.2",
    "rich>=13.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
    "ruff>=0.5",
]

[project.scripts]
rebalancer = "rebalancer.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
ignore = ["E501"]
